auth 완성 및 jwt 발급받아 authorized 받고 게시글 ins까지 완료

+++ src/main/java/com/asdf/minilog/service/ArticleService.java
@@ -4,6 +4,7 @@ import com.asdf.minilog.dto.ArticleResponseDto;
 import com.asdf.minilog.entity.Article;
 import com.asdf.minilog.entity.User;
 import com.asdf.minilog.exception.ArticleNotFoundException;
+import com.asdf.minilog.exception.NotAuthorizedException;
 import com.asdf.minilog.exception.UserNotFoundException;
 import com.asdf.minilog.repository.ArticleRepository;
 import com.asdf.minilog.repository.UserRepository;
@@ -42,7 +43,7 @@ public class ArticleService {
         return EntityDtoMapper.toDto(savedArticle);
     }
 
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long authorId, Long articleId) {
         Article article =
 
                 articleRepository
@@ -52,10 +53,15 @@ public class ArticleService {
                                         new ArticleNotFoundException(
                                                 String.format("해당  아이디(%d)를 가진 게시글을 찾을 수 " +
                                                         "없습니다.", articleId)));
+        if(!article.getAuthor().getId().equals(authorId)) {
+            throw new NotAuthorizedException("게시글 작성자만 삭제할 수 있습니다.");
+        }
+
+
         articleRepository.deleteById(articleId);
     }
 
-    public ArticleResponseDto updateArticle(Long articleId, String content) {
+    public ArticleResponseDto updateArticle(Long authorId, Long articleId, String content) {
         Article article =
                 articleRepository
                         .findById(articleId)
@@ -63,6 +69,10 @@ public class ArticleService {
                                 () ->
                                         new ArticleNotFoundException(
                                                 String.format("해당 아이디(%d)를 찾을 수 없습니다.", articleId)));
+
+        if(!article.getAuthor().getId().equals(authorId)) {
+            throw new NotAuthorizedException("게시글 작성자만 수정할 수 있습니다.");
+        }
 
         article.setContent(content);
         Article updatedArticle =  articleRepository.save(article);
